Use bit-vec instead of bitvec_simd for now
The problem of bitvec_simd is that it uses smallvec for its internal
storage. That is really bad for us, because if we have more than 1024
frames (which is usually the case for real applications) SmallVec will
spill its data onto the heap and then the size of BitVec (224bytes =
1024/8 + 8 + padding) are unused (1024/8 from the SmallVec, 8 from
additional members in BitVec). The problem for us is that we are going
to have a lot of those BitVec(s) since 2 BitVec(s) are needed for one
qubit.