Adapt the frame popping logic
Before we could only pop if the number of the x and z corrections on the
qubit where the same. That is too strict for us (e.g., when using
something like the move_z_to_z method), so I replaced "pop" with
"pop_or_false", which defaults to false if necessary.