Refactor the tracker API
Originally I thought it would be enough to have one type "Frames" that
would be generic over different "Storage" types. This works well as long
as the "Storages" all have the same underlining structure of storing the
Paulis per qubit, e.g., PauliVec in our case. However, when this
structure is different, things would get really ugly. One could
introduce something like a "Stack" trait that defines the required
interface for the structure, but this would require a lot of boilerplate
and makes things more complicated than necessary. Instead I put the
characteristic tracking methods from Frames into a trait (Tracker). This
is more flexible and makes things clearer. The disadvantage is that we
have to duplicate the tracking logic for different Tracker
implementors, however, this logic is actually very simple; the more
difficult part is to apply the logic onto the underlining data
structure, and that is simpler with the new trait.