Implement a live tracker + minor changes
First implementation of a live tracker (this implementation is the
reason why we performed the refactor in the last commit). The live
tracker doesn't keep track of all frames, but instead updates one single
frame. This can be used when the quantum computation is actually
performed and we are only interested in total corrections and not where
they came from. Specifically one would update the tracker depending on
the real measurement results.
The minor changes just include, clean-ups and minor refactors.