Make the circuit module public
Originally, I wanted to make much later public, since I thought it would
be very unstable. However, now I think that it actually may be not that
unstable (let's hope that will be true). Making it public has the main
benefit that I can write better examples in the documentation and
README.md (it's just more intuitive to understand the Pauli tracking, if
it is connected to a real circuit).
I also removed some code in this module which was pretty unnecessary for
this crate.