Refactor PauliVec
PauliVec had a fixed inner storage type (a bit-vector). Now it is
generic over the inner storage type, which has to implement the
BoolVector trait. The main reason why I did this is that there is no
best bit-vector implementation, e.g., implementations that use SIMD
stuff have the benefit that operations like xor, and, ... are very fast,
however, read and write operations are usually slower. What is best,
depends on the circuit structure. Also, there are many librararies that
implement bit-vectors and I don't want to restrict the code to just one
of those.
The code has to be cleaned up.