Simplify the time graph
Previously the TimeGraph had the dependencies layered, because I thought
that that could be used for more aggressive break conditions when doing
the sweeps. However, this is not true, at least not in the form how it
was implemented. One would need to consider the edges between the layers
and update them when a node gets measured, or alternatively move the
nodes through the layers. Both methods have a serious overhead, since it
requires looping through the whole graph for each measurement. Since I'm
not sure whether the performance gain by the more aggressive break
conditions would compensate that, and since it definitely is an overhead
when we don't want to break the sweeps, I completely removed the
layering in the TimeGraph. Maybe there will we a layered version in
future, if it is worth it, but currently I don't think so.
I also cleaned up some debugging remnants.