Refactor the StackStorage trait
*Note the doc examples are currently breaking*

I introduced a new trait Collections which is basically like
StackStorage, however, instead of having an associated BooleanVector
type, it is generic. This has the benefit that we can use for the
Live(Vector) pauli tracker. The disadvantage is that there's an
additional level of generics, which makes the types more complex.